refactor: move Store to a separate module
